requirements.txt
flask==1.0.2
sqlalchemy==1.3.3
python-dotenv
mysqlclient
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy